requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
prometheus-client==0.19.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
matplotlib==3.8.2
seaborn==0.13.0
pydantic-settings==2.1.0
